server/requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_cors
python-dotenv
flask-jwt-extended